requirements.txt
phidata
requests
duckduckgo-search
python-dotenv
groq